squash

+++ tests/llmcompressor/transformers/gptq/test_gptq_oneshot.py
@@ -181,9 +181,7 @@ def test_oneshot_application(recipe, tmp_path):
     device = "cuda:0" if torch.cuda.is_available() else "cpu"
 
     # Load original model for numerical comparison
-    original_model = AutoModelForCausalLM.from_pretrained(
-        model_id, torch_dtype=torch.float16, device_map=device
-    )
+    original_model = AutoModelForCausalLM.from_pretrained(model_id, device_map=device)
     tokenizer = AutoTokenizer.from_pretrained(model_id)
 
     # Create test input
